Gulp tasks extracted to separate files
